+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/storage/portablegrid/PortableGridBlockEntityTicker.java
@@ -21,6 +21,6 @@ class PortableGridBlockEntityTicker extends AbstractBlockEntityTicker<AbstractPo
                      final BlockPos pos,
                      final BlockState state,
                      final AbstractPortableGridBlockEntity blockEntity) {
-        blockEntity.updateDiskStateIfNecessaryInLevel();
+        blockEntity.update(state);
     }
 }